requirements.txt
streamlit==1.34.0
PyPDF2==3.0.1
pdfplumber==0.10.2
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
pillow==10.1.0
python-magic==0.4.27
numpy==1.26.3
google-generativeai==0.4.0
spacy==3.7.2
blis==0.7.9
thinc==8.2.2
cymem==2.0.8
preshed==3.0.9
wasabi==1.1.2
srsly==2.4.8
catalogue==2.0.10
